FUNCIONANDO TODO OK 16-05

+++ base/views.py
@@ -1,7 +1,60 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.http import HttpResponse
+
+
+def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "Usuario inexistente")
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "El Usuario o la contraseña no existe")
+
+    context = {"page": page}
+
+    return render(request, "base/login_register.html", context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect("home")
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Hubo un error durante el registro")
+    context = {"form": form}
+    return render(request, "base/login_register.html", context)
 
 
 def home(request):
@@ -9,6 +62,7 @@ def home(request):
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q)
     )
+
     topics = Topic.objects.all()
     room_count = rooms.count()
     context = {"rooms": rooms, "topics": topics, "room_count": room_count}
@@ -17,13 +71,22 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {"room": room}
+    room_messages = room.message_set.all().order_by("-created")
 
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user, room=room, body=request.POST.get("body")
+        )
+        return redirect("room", pk=room.id)
+
+    context = {"room": room, "room_messages": room_messages}
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="/login")
 def createRoom(request):
     form = RoomForm()
+
     if request.method == "POST":
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -34,11 +97,17 @@ def createRoom(request):
     return render(request, "base/rooms_form.html", context)
 
 
+@login_required(login_url="/login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("No puedes editar un room no crado por ti")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
+
         if form.is_valid():
             form.save()
             return redirect("home")
@@ -47,9 +116,15 @@ def updateRoom(request, pk):
     return render(request, "base/rooms_form.html", context)
 
 
+@login_required(login_url="/login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("No puedes eliminar un room no crado por ti")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
+
     return render(request, "base/delete.html", {"obj": room})
